Export mock_context()
