+++ runtime/test/src/test.rs
@@ -19,7 +19,7 @@ use std::str::FromStr;
 use test_store::STORE;
 use web3::types::H160;
 
-use crate::common::mock_data_source;
+use crate::common::{mock_data_source, mock_abi};
 
 mod abi;
 
@@ -119,70 +119,6 @@ fn test_module(
     api_version: Version,
 ) -> WasmInstance<Chain> {
     test_valid_module_and_store(subgraph_id, data_source, api_version).0
-}
-
-fn mock_host_exports(
-    subgraph_id: DeploymentHash,
-    data_source: DataSource,
-    store: Arc<impl SubgraphStore>,
-    api_version: Version,
-) -> HostExports<Chain> {
-    let templates = vec![DataSourceTemplate {
-        kind: String::from("ethereum/contract"),
-        name: String::from("example template"),
-        network: Some(String::from("mainnet")),
-        source: TemplateSource {
-            abi: String::from("foo"),
-        },
-        mapping: Mapping {
-            kind: String::from("ethereum/events"),
-            api_version,
-            language: String::from("wasm/assemblyscript"),
-            entities: vec![],
-            abis: vec![],
-            event_handlers: vec![],
-            call_handlers: vec![],
-            block_handlers: vec![],
-            link: Link {
-                link: "link".to_owned(),
-            },
-            runtime: Arc::new(vec![]),
-        },
-    }];
-
-    let network = data_source.network.clone().unwrap();
-    HostExports::new(
-        subgraph_id,
-        &data_source,
-        network,
-        Arc::new(templates),
-        Arc::new(graph_core::LinkResolver::from(IpfsClient::localhost())),
-        store,
-    )
-}
-
-fn mock_context(
-    deployment: DeploymentLocator,
-    data_source: DataSource,
-    store: Arc<impl SubgraphStore>,
-    api_version: Version,
-) -> MappingContext<Chain> {
-    MappingContext {
-        logger: test_store::LOGGER.clone(),
-        block_ptr: BlockPtr {
-            hash: Default::default(),
-            number: 0,
-        },
-        host_exports: Arc::new(mock_host_exports(
-            deployment.hash.clone(),
-            data_source,
-            store.clone(),
-            api_version,
-        )),
-        state: BlockState::new(store.writable(&deployment).unwrap(), Default::default()),
-        proof_of_indexing: None,
-        host_fns: Arc::new(Vec::new()),
-    }
 }
 
 trait WasmInstanceExt {
